requirements-multimodal.txt
# Multimodal RAG System Requirements

## Core Dependencies
sentence-transformers>=2.2.0
transformers>=4.35.0
torch>=2.0.0
pillow>=10.0.0
numpy>=1.24.0

## Document Processing
PyMuPDF>=1.23.0  # PDF processing
python-docx>=1.0.0  # Word documents (optional)

## Audio Processing
openai-whisper>=20231117
librosa>=0.10.0
soundfile>=0.12.0

## Vector Database
qdrant-client>=1.7.0

## API Service
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

## CLI and Evaluation
click>=8.1.0
pandas>=2.0.0
openai>=1.0.0  # For LLM judge

## Existing Namel3ss Dependencies
# (already in requirements.txt)
# aiohttp
# jinja2
# pydantic
# etc.

## Optional: Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
mypy>=1.0.0
